requirements.txt
scikit-image
numpy
pillow
pure-python-adb